New utils file for common helper functions
